py-pdf-text-converter/requirements.txt
streamlit
PyMuPDF
pdfminer.six
pytesseract
pdf2image
pillow